+++ bmc-managementdashboard/src/main/java/com/oracle/bmc/managementdashboard/model/CreateManagementDashboardDetails.java
@@ -197,6 +197,15 @@ public class CreateManagementDashboardDetails {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("drilldownConfig")
+        private java.util.List<Object> drilldownConfig;
+
+        public Builder drilldownConfig(java.util.List<Object> drilldownConfig) {
+            this.drilldownConfig = drilldownConfig;
+            this.__explicitlySet__.add("drilldownConfig");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("freeformTags")
         private java.util.Map<String, String> freeformTags;
 
@@ -241,6 +250,7 @@ public class CreateManagementDashboardDetails {
                             type,
                             isFavorite,
                             parametersConfig,
+                            drilldownConfig,
                             freeformTags,
                             definedTags);
             __instance__.__explicitlySet__.addAll(__explicitlySet__);
@@ -269,6 +279,7 @@ public class CreateManagementDashboardDetails {
                             .type(o.getType())
                             .isFavorite(o.getIsFavorite())
                             .parametersConfig(o.getParametersConfig())
+                            .drilldownConfig(o.getDrilldownConfig())
                             .freeformTags(o.getFreeformTags())
                             .definedTags(o.getDefinedTags());
 
@@ -397,6 +408,12 @@ public class CreateManagementDashboardDetails {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("parametersConfig")
     java.util.List<Object> parametersConfig;
+
+    /**
+     * Drill-down configuration to define the destination of a drill-down action.
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("drilldownConfig")
+    java.util.List<Object> drilldownConfig;
 
     /**
      * Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.