+++ bmc-managementdashboard/src/main/java/com/oracle/bmc/managementdashboard/model/ManagementDashboard.java
@@ -260,6 +260,15 @@ public class ManagementDashboard {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("drilldownConfig")
+        private java.util.List<Object> drilldownConfig;
+
+        public Builder drilldownConfig(java.util.List<Object> drilldownConfig) {
+            this.drilldownConfig = drilldownConfig;
+            this.__explicitlySet__.add("drilldownConfig");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("freeformTags")
         private java.util.Map<String, String> freeformTags;
 
@@ -311,6 +320,7 @@ public class ManagementDashboard {
                             savedSearches,
                             lifecycleState,
                             parametersConfig,
+                            drilldownConfig,
                             freeformTags,
                             definedTags);
             __instance__.__explicitlySet__.addAll(__explicitlySet__);
@@ -346,6 +356,7 @@ public class ManagementDashboard {
                             .savedSearches(o.getSavedSearches())
                             .lifecycleState(o.getLifecycleState())
                             .parametersConfig(o.getParametersConfig())
+                            .drilldownConfig(o.getDrilldownConfig())
                             .freeformTags(o.getFreeformTags())
                             .definedTags(o.getDefinedTags());
 
@@ -516,6 +527,12 @@ public class ManagementDashboard {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("parametersConfig")
     java.util.List<Object> parametersConfig;
+
+    /**
+     * Drill-down configuration to define the destination of a drill-down action.
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("drilldownConfig")
+    java.util.List<Object> drilldownConfig;
 
     /**
      * Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.